Moves the userToDatabase.js to the helper folder and updates the reference to it in infor.js
