+++ src/helpers/userToDatabase.js
