+++ src/routes/info.js
@@ -2,7 +2,7 @@ const fs = require('fs');
 const path = require('path');
 const YotiClient = require('yoti-node-sdk');
 const dbConn = require('../database/db_connection');
-const yotiToDb = require('./userToDatabase.js');
+const yotiToDb = require('../helpers/userToDatabase.js');
 
 const CLIENT_SDK_ID = "8a4dcb2a-9ed6-4d44-9a55-12b581bb5e64";
 const PEM = fs.readFileSync(path.join(__dirname, '../../keys/help-access-security.pem'));
@@ -35,7 +35,7 @@ module.exports = [{
           console.log("There was an error adding user: ", err);
         }
         reply.view('info', context);
-      });      
+      });
     }).catch((err) => {
       console.error(err);
       reply.view('error', {
